returns a list of files provided the search criteria

returns a list of files provided the search criteria

raw.findFiles(
  pfad,
  project = ".*",
  date = ".*",
  user = ".*",
  instrument = ".*",
  sample = ".*",
  md5 = "",
  fullPath = TRUE,
  underscoreComments = TRUE
)

Arguments

pfad
path of RAW files

project
name of project

date
date in format YYYYMMDD

user
2 initials for user

instrument
name of instrument, such as 'vsm'

sample
sample name

md5
single string with comma separated abbreviated MD5 sums

fullPath
if TRUE, returns full path, otherwise filename only

underscoreComments
if TRUE, comments can have underscores

Value

list with filenames with certain checksum

Examples

pfad = raw.getSamplePath()
file.list = raw.findFiles(pfad, date='201606')  # all files from Jan 2019
print(file.list)
#> [1] "/private/var/folders/n7/c8bjcf7n25g74wqr3_7cgpvh0000gt/T/RtmpxwBBwh/temp_libpath28f77071045d/RAWdataR/extdata/20160607_Optical_AN_Spectrophotometer_NM20151109CuPc100gl150-A2(UDS).txt"
md5String = raw.getMD5str(file.list)
file.list = raw.findFiles(pfad, md5 = md5String)
print(file.list)
#> [1] "/private/var/folders/n7/c8bjcf7n25g74wqr3_7cgpvh0000gt/T/RtmpxwBBwh/temp_libpath28f77071045d/RAWdataR/extdata/20141215_FePc_TG_vsm_Tb33Al66_20141215-Tb33Al66-ac1Oe-20Hz.txt"          
#> [2] "/private/var/folders/n7/c8bjcf7n25g74wqr3_7cgpvh0000gt/T/RtmpxwBBwh/temp_libpath28f77071045d/RAWdataR/extdata/20160607_Optical_AN_Spectrophotometer_NM20151109CuPc100gl150-A2(UDS).txt"